Added 'current step' and 'decrement button'
This probably should have been two changelists. Now we can grab the
current step from an array of steps and say “BOIL” and start time for
boil. Right now this is hard coded in store.js because I don’t know how
to data. Also I made a button to decrement the timer!